add file input_types.py for graphql mutations.
