+++ apps/etl/mutations.py
@@ -1,5 +1,4 @@
 import logging
-from typing import List
 
 import strawberry
 from asgiref.sync import sync_to_async
@@ -10,6 +9,7 @@ from apps.etl.extraction.sources.gidd.extract import GIDDExtraction
 from apps.etl.extraction.sources.glide.extract import GlideExtraction
 from apps.etl.extraction.sources.idu.extract import IDUExtraction
 from apps.etl.extraction.sources.ifrc_event.extract import IFRCEventExtraction
+from apps.etl.input_types import PipelineRetriggerInput, TransformRetriggerInput
 from apps.etl.models import ExtractionData, Transform
 from apps.etl.transform.sources.emdat import EMDATTransformHandler
 from apps.etl.transform.sources.gidd import GIDDTransformHandler
@@ -34,16 +34,6 @@ source_transform_map = {
     ExtractionData.Source.IDU: IDUTransformHandler,
     ExtractionData.Source.DREF: IFRCEventTransformHandler,
 }
-
-
-@strawberry.input
-class PipelineRetriggerInput:
-    trace_id: List[int]
-
-
-@strawberry.input
-class TransformRetriggerInput:
-    transform_id: List[int]
 
 
 @strawberry.type
